ml/requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
